+++ R/ts.R
@@ -175,17 +175,25 @@ bic.climex.fit.gev <- bic.climex.fit.gpd <- bic.default <- function( x ){
 ##' @param x Either a \pkg{xts} class object or a list of those.
 ##' @param time.unit Minimal differences found when applying the
 ##'   \code{\link[base]{diff}} function to \strong{x}. Default = 1.
+##' @param mc.cores A numerical input specifying the number of cores
+##'   to use for the multi core application of the function (see
+##'   \code{\link[parallel]{detectCores}}). This functionality is only
+##'   available if the input is a list of different objects. If NULL,
+##'   the function will be calculated classically with only one core.
+##'   Default = NULL.
 ##'
 ##' @seealso \code{\link{remove.seasonality}}
 ##'
 ##' @family ts
 ##' @export
 ##' @importFrom xts xts
+##' @importFrom parallel mclapply
 ##' @import lubridate
 ##' 
 ##' @return Object of the same class as the input \strong{x}.
 ##' @author Philipp Mueller
-remove.incomplete.years <- function( x, time.unit = 1 ){
+remove.incomplete.years <- function( x, time.unit = 1,
+                                    mc.cores = NULL ){
   UseMethod( "remove.incomplete.years" )
 }
 ##' @title Remove incomplete years
@@ -202,19 +210,32 @@ remove.incomplete.years <- function( x, time.unit = 1 ){
 ##' @param x A list of \pkg{xts} class objects.
 ##' @param time.unit Minimal differences found when applying the
 ##'   \code{\link[base]{diff}} function to \strong{x}. Default = 1.
+##' @param mc.cores A numerical input specifying the number of cores
+##'   to use for the multi core application of the function (see
+##'   \code{\link[parallel]{detectCores}}). This functionality is only
+##'   available if the input is a list of different objects. If NULL,
+##'   the function will be calculated classically with only one core.
+##'   Default = NULL.
 ##'
 ##' @seealso \code{\link{remove.seasonality}}
 ##'
 ##' @family ts
 ##' @export
 ##' @importFrom xts xts
+##' @importFrom parallel mclapply
 ##' @import lubridate
 ##' 
 ##' @return Object of the same class as the input \strong{x}.
 ##' @author Philipp Mueller
-remove.incomplete.years.list <- function( x, time.unit = 1 ){
-  return( lapply( x, remove.incomplete.years,
-                 time.unit = time.unit ) )
+remove.incomplete.years.list <- function( x, time.unit = 1,
+                                         mc.cores = NULL ){
+  if ( !is.null( mc.cores ) ){
+    return( mclapply( x, remove.incomplete.years,
+                     time.unit = time.unit, mc.cores = mc.cores ) )
+  } else {
+    return( lapply( x, remove.incomplete.years,
+                   time.unit = time.unit, mc.cores = mc.cores ) )
+  }
 }
 ##' @title Remove incomplete years
 ##' @description Removes all years, which contain either a NA or are
@@ -230,18 +251,25 @@ remove.incomplete.years.list <- function( x, time.unit = 1 ){
 ##' @param x A \pkg{xts} class object.
 ##' @param time.unit Minimal differences found when applying the
 ##'   \code{\link[base]{diff}} function to \strong{x}. Default = 1.
+##' @param mc.cores A numerical input specifying the number of cores
+##'   to use for the multi core application of the function (see
+##'   \code{\link[parallel]{detectCores}}). This functionality is only
+##'   available if the input is a list of different objects. If NULL,
+##'   the function will be calculated classically with only one core.
+##'   Default = NULL.
 ##'
 ##' @seealso \code{\link{remove.seasonality}}
 ##'
 ##' @family ts
 ##' @export
 ##' @importFrom xts xts
+##' @importFrom parallel mclapply
 ##' @import lubridate
 ##' 
 ##' @return Object of the same class as the input \strong{x}.
 ##' @author Philipp Mueller
 remove.incomplete.years.xts <- remove.incomplete.years.default <-
-  function( x, time.unit = 1 ){
+  function( x, time.unit = 1, mc.cores = NULL ){
     if ( !any( class( x ) == "xts" ) )
       stop(
           "Provided element of the wrong class in remove.incomplete.years. Provide object of class xts instead!" )
@@ -332,8 +360,16 @@ remove.incomplete.years.xts <- remove.incomplete.years.default <-
 ##'   them.
 ##' @param number.of.years The minimum number of complete years in the
 ##'   series. Default = 30.
+##' @param mc.cores A numerical input specifying the number of cores
+##'   to use for the multi core application of the function (see
+##'   \code{\link[parallel]{detectCores}}). This functionality is only
+##'   available if the input is a list of different objects. If NULL,
+##'   the function will be calculated classically with only one core.
+##'   Default = NULL.
 ##'
 ##' @export
+##' 
+##' @importFrom parallel mclapply
 ##'
 ##' @family ts
 ##'
@@ -343,7 +379,8 @@ remove.incomplete.years.xts <- remove.incomplete.years.default <-
 ##'   containing only those elements, which indeed have more than N
 ##'   years of complete data.
 ##' @author Philipp Mueller 
-check.completeness <- function( x, number.of.years = 30 ){
+check.completeness <- function( x, number.of.years = 30,
+                               mc.cores = NULL){
   UseMethod( "check.completeness" )
 }
 ##' @title Checking for N complete years of data
@@ -354,18 +391,35 @@ check.completeness <- function( x, number.of.years = 30 ){
 ##' @param x A list of \pkg{xts} class objects.
 ##' @param number.of.years The minimum number of complete years in the
 ##'   series. Default = 30.
+##' @param mc.cores A numerical input specifying the number of cores
+##'   to use for the multi core application of the function (see
+##'   \code{\link[parallel]{detectCores}}). This functionality is only
+##'   available if the input is a list of different objects. If NULL,
+##'   the function will be calculated classically with only one core.
+##'   Default = NULL.
 ##'
 ##' @export
+##'
+##' @importFrom parallel mclapply
 ##'
 ##' @return Trimmed list containing only those elements, which indeed
 ##'   have more than N years of complete data.
 ##' @author Philipp Mueller
-check.completeness.list <- function( x, number.of.years = 30 ){
-  ## Apply the checks to all elements of the list.
-  x.list.check <-
-    Reduce( c, lapply( x, check.completeness.xts,
-                      number.of.years = number.of.years ) )
-
+check.completeness.list <- function( x, number.of.years = 30,
+                                    mc.cores = NULL ){
+  if ( !is.null( mc.cores ) ){
+    ## Apply the checks to all elements of the list.
+    x.list.check <-
+      Reduce( c, mclapply( x, check.completeness.xts,
+                          number.of.years = number.of.years,
+                          mc.cores = mc.cores ) )
+  } else {
+    ## Apply the checks to all elements of the list.
+    x.list.check <-
+      Reduce( c, lapply( x, check.completeness.xts,
+                        number.of.years = number.of.years,
+                        mc.cores = mc.cores ) )
+  }
   ## Return only those element matching the conditions
   return( x[ x.list.check ] )
 }
@@ -380,8 +434,15 @@ check.completeness.list <- function( x, number.of.years = 30 ){
 ##' @param x A time series of class \pkg{xts}.
 ##' @param number.of.years The minimum number of complete years in the
 ##'   series. Default = 30.
+##' @param mc.cores A numerical input specifying the number of cores
+##'   to use for the multi core application of the function (see
+##'   \code{\link[parallel]{detectCores}}). This functionality is only
+##'   available if the input is a list of different objects. If NULL,
+##'   the function will be calculated classically with only one core.
+##'   Default = NULL.
 ##'
 ##' @importFrom xts is.xts
+##' @importFrom parallel mclapply
 ##'
 ##' @export
 ##' 
@@ -390,7 +451,7 @@ check.completeness.list <- function( x, number.of.years = 30 ){
 ##' @return Logical
 ##' @author Philipp Mueller 
 check.completeness.xts <- check.completeness.default <-
-  function( x, number.of.years = 30 ){
+  function( x, number.of.years = 30, mc.cores = NULL ){
     ## Only accept objects of class 'xts'
     if ( !is.xts( x ) ){
       stop( "Only objects of class 'xts' accepted!" )
@@ -430,14 +491,21 @@ check.completeness.xts <- check.completeness.default <-
 ##'   objects.
 ##'
 ##' @param x Either an object of class \pkg{xts} or a list of those.
+##' @param mc.cores A numerical input specifying the number of cores
+##'   to use for the multi core application of the function (see
+##'   \code{\link[parallel]{detectCores}}). This functionality is only
+##'   available if the input is a list of different objects. If NULL,
+##'   the function will be calculated classically with only one core.
+##'   Default = NULL.
 ##'
 ##' @family ts
 ##'
 ##' @importFrom xts xts
+##' @importFrom parallel mclapply
 ##' @import lubridate
 ##' @return Some class as input.
 ##' @author Philipp Mueller
-remove.seasonality <- function( x ){
+remove.seasonality <- function( x, mc.cores = NULL ){
   UseMethod( "remove.seasonality" )
 }
 ##' @title Remove seasonality
@@ -460,15 +528,26 @@ remove.seasonality <- function( x ){
 ##'   objects.
 ##'
 ##' @param x A list of objects of class \pkg{xts}.
+##' @param mc.cores A numerical input specifying the number of cores
+##'   to use for the multi core application of the function (see
+##'   \code{\link[parallel]{detectCores}}). This functionality is only
+##'   available if the input is a list of different objects. If NULL,
+##'   the function will be calculated classically with only one core.
+##'   Default = NULL.
 ##'
 ##' @family ts
 ##'
 ##' @importFrom xts xts
+##' @importFrom parallel mclapply
 ##' @import lubridate
 ##' @return Some class as input.
 ##' @author Philipp Mueller
-remove.seasonality.list <- function( x ){
-  return( lapply( x, remove.seasonality ) )
+remove.seasonality.list <- function( x, mc.cores = NULL ){
+  if ( !is.null( mc.cores ) ){
+    return( mclapply( x, remove.seasonality, mc.cores = mc.cores ) )
+  } else {
+    return( lapply( x, remove.seasonality, mc.cores = mc.cores ) )
+  }
 }
 ##' @title Remove seasonality
 ##' @description Calculates the seasonal component of a time series
@@ -490,111 +569,121 @@ remove.seasonality.list <- function( x ){
 ##'   objects.
 ##'
 ##' @param x An object of class \pkg{xts}.
+##' @param mc.cores A numerical input specifying the number of cores
+##'   to use for the multi core application of the function (see
+##'   \code{\link[parallel]{detectCores}}). This functionality is only
+##'   available if the input is a list of different objects. If NULL,
+##'   the function will be calculated classically with only one core.
+##'   Default = NULL.
 ##'
 ##' @family ts
 ##'
 ##' @importFrom xts xts
+##' @importFrom parallel mclapply
 ##' @import lubridate
 ##' @return Some class as input.
 ##' @author Philipp Mueller
-remove.seasonality.xts <- remove.seasonality.default <- function( x ){
-  if ( !any( class( x ) == "xts" ) )
-    stop( "Provided element of the wrong class in remove.seasonality. Provide object of class xts instead!" )  
-  ## To ensure reasonal results 
-  if ( min( x, na.rm = TRUE ) == -999 )
-    x[ x == -999 ] <- NA
-  x.clean <- remove.incomplete.years( x )
-  
-  ## When e.g. no maximal temperatures are provided for a station
-  ## the function will exist
-  if ( length( x.clean ) == 0 )
-    return( list( x = NA, modified = FALSE ) )
-  
-  ## Was time series modified?
-  ifelse( length( x ) == length( x.clean ),
-         MODIFIED <- FALSE, MODIFIED <- TRUE )
-  
-  ## maybe to much data was remove
-  if ( length( x ) < 365*5 ){
-    warning( "After the removal of the incomplete years less than five years are remaining in the data set in remove.seasonality" )
-    return( list( x = NA, modified = MODIFIED ) )
-  }
-  
-  ## Seasonal component of the time series
-  x.seas <- stats::stl( stats::ts( as.numeric( x.clean ),
-                                  frequency = 365 ),
-                       s.window = 12 )$time.series[ , 1 ]
-  
-  ## Usually the seasonal term is more than just a constant sinusoidal
-  ## function but exhibits some kind of trend itself. That's why it's
-  ## only logical to use the same or the nearest year in the seasonal
-  ## component for subtracting of original series.
-  ## Subtracting all years which are also present in the seasonal
-  ## component
-  x.clean.years <- unique( year( x.clean ) )
-  x.years <- unique( year( x ) )
-  for ( yy in x.clean.years )
-    x[ year( x ) == yy ] <- x[ year( x ) == yy ] -
-      x.seas[ which( year( x.clean ) == yy ) ]
-  
-  ## Looping over all years don't accounted yet and substracting the
-  ## year nearst in x.seas to the original one. The alignment of the
-  ## to data sets has to be handled with care!
-  for ( yy in x.years[ !x.years %in% x.clean.years ] ){
-    ## closest seasonal component; Caution, can be a vector!
-    ## if the year in x is a leap year one has to search for the
-    ## closest leap year.
-    ## If the closest year has not 366 days its distance is set to 999
-    if ( length( x[ year( x ) == yy ] ) == 366 ){
-      closest.year.vec <-  abs( x.clean.years - yy )
-      for ( cc in 1 : length( closest.year.vec ) ){
-        if ( length( x.clean[
-            year( x.clean ) == x.clean.years[
-                                   which( closest.year.vec ==
-                                          min( closest.year.vec )
-                                         ) ] ] )
-            == 366 ){
-          closest.year.pos <- which( closest.year.vec ==
-                                     min( closest.year.vec ) )
-          break
-        } else
-          closest.year.vec[ which( closest.year.vec ==
-                                   min( closest.year.vec ) ) ] <- 999
-      }
-    } else {
-      closest.year.pos <- which( abs( x.clean.years - yy ) ==
-                                 min( abs( x.clean.years - yy ) ) )
+remove.seasonality.xts <- remove.seasonality.default <-
+  function( x, mc.cores = NULL ){
+    if ( !any( class( x ) == "xts" ) )
+      stop( "Provided element of the wrong class in remove.seasonality. Provide object of class xts instead!" )  
+    ## To ensure reasonal results 
+    if ( min( x, na.rm = TRUE ) == -999 )
+      x[ x == -999 ] <- NA
+    x.clean <- remove.incomplete.years( x )
+    
+    ## When e.g. no maximal temperatures are provided for a station
+    ## the function will exist
+    if ( length( x.clean ) == 0 )
+      return( list( x = NA, modified = FALSE ) )
+    
+    ## Was time series modified?
+    ifelse( length( x ) == length( x.clean ),
+           MODIFIED <- FALSE, MODIFIED <- TRUE )
+    
+    ## maybe to much data was remove
+    if ( length( x ) < 365*5 ){
+      warning( "After the removal of the incomplete years less than five years are remaining in the data set in remove.seasonality" )
+      return( list( x = NA, modified = MODIFIED ) )
     }
-    closest.year <- x.clean.years[ closest.year.pos[ 1 ] ] 
     
-    ## subtracts only those days of the closest year in x.seas which
-    ## are also present in the x time series. ydays give the calender
-    ## date (1--366)
-    days.clean <- yday( x.clean[ year( x.clean ) == closest.year ] )
-    days.x <- yday( x[ year( x ) == yy ] )
-    ## Even if the length of x is not 366 it can still be a leap year
-    ## and the 31.12. will be of the value 366. This way the seasonal
-    ## part might not has the some number of entries (the closest year
-    ## is not a leap year)
-    if ( min( days.x ) == 1 && max( days.x ) == 366 &&
-         length( days.x ) < 366 ){
-      ## Somewhere at least one day is missing. This "step" has to
-      ## be detected and the all days after the step have to be
-      ## reduced by one
-      step <- which( days.x - c( 1 : length( days.x ) ) > 0 )[ 1 ]
-      days.x[ step : length( days.x ) ] <- days.x[
-          step : length( days.x ) ] - 1
-    } else if ( max( days.x ) == 366 && max( days.clean ) != 366 )
-      days.x <- days.x - 1
+    ## Seasonal component of the time series
+    x.seas <- stats::stl( stats::ts( as.numeric( x.clean ),
+                                    frequency = 365 ),
+                         s.window = 12 )$time.series[ , 1 ]
     
-    if ( min( days.x ) == 0 )
-      stop( "Wait. Something went wrong in remove.seasonality. Check for the correct alignment of the seasonal part and the original time series!" )
-    x[ year( x ) == yy ] <- x[ year( x ) == yy ] - 
-      as.numeric( x.seas[ year( x.clean ) == closest.year ][
-          days.clean %in% days.x ] )
+    ## Usually the seasonal term is more than just a constant
+    ## sinusoidal function but exhibits some kind of trend
+    ## itself. That's why it's only logical to use the same or the
+    ## nearest year in the seasonal component for subtracting of
+    ## original series. Subtracting all years which are also present
+    ## in the seasonal component
+    x.clean.years <- unique( year( x.clean ) )
+    x.years <- unique( year( x ) )
+    for ( yy in x.clean.years )
+      x[ year( x ) == yy ] <- x[ year( x ) == yy ] -
+        x.seas[ which( year( x.clean ) == yy ) ]
+    
+    ## Looping over all years don't accounted yet and substracting the
+    ## year nearst in x.seas to the original one. The alignment of the
+    ## to data sets has to be handled with care!
+    for ( yy in x.years[ !x.years %in% x.clean.years ] ){
+      ## closest seasonal component; Caution, can be a vector!
+      ## if the year in x is a leap year one has to search for the
+      ## closest leap year.
+      ## If the closest year has not 366 days its distance is set to
+      ## 999
+      if ( length( x[ year( x ) == yy ] ) == 366 ){
+        closest.year.vec <-  abs( x.clean.years - yy )
+        for ( cc in 1 : length( closest.year.vec ) ){
+          if ( length( x.clean[
+              year( x.clean ) == x.clean.years[
+                                     which( closest.year.vec ==
+                                            min( closest.year.vec )
+                                           ) ] ] )
+              == 366 ){
+            closest.year.pos <- which( closest.year.vec ==
+                                       min( closest.year.vec ) )
+            break
+          } else
+            closest.year.vec[
+                which( closest.year.vec ==
+                       min( closest.year.vec ) ) ] <- 999
+        }
+      } else {
+        closest.year.pos <- which( abs( x.clean.years - yy ) ==
+                                   min( abs( x.clean.years - yy ) ) )
+      }
+      closest.year <- x.clean.years[ closest.year.pos[ 1 ] ] 
+      
+      ## subtracts only those days of the closest year in x.seas which
+      ## are also present in the x time series. ydays give the
+      ## calender date (1--366)
+      days.clean <- yday( x.clean[ year( x.clean ) == closest.year ] )
+      days.x <- yday( x[ year( x ) == yy ] )
+      ## Even if the length of x is not 366 it can still be a leap
+      ## year and the 31.12. will be of the value 366. This way the
+      ## seasonal part might not has the some number of entries (the
+      ## closest year is not a leap year)
+      if ( min( days.x ) == 1 && max( days.x ) == 366 &&
+           length( days.x ) < 366 ){
+        ## Somewhere at least one day is missing. This "step" has to
+        ## be detected and the all days after the step have to be
+        ## reduced by one
+        step <- which( days.x - c( 1 : length( days.x ) ) > 0 )[ 1 ]
+        days.x[ step : length( days.x ) ] <- days.x[
+            step : length( days.x ) ] - 1
+      } else if ( max( days.x ) == 366 && max( days.clean ) != 366 )
+        days.x <- days.x - 1
+      
+      if ( min( days.x ) == 0 )
+        stop( "Wait. Something went wrong in remove.seasonality. Check for the correct alignment of the seasonal part and the original time series!" )
+      x[ year( x ) == yy ] <- x[ year( x ) == yy ] - 
+        as.numeric( x.seas[ year( x.clean ) == closest.year ][
+            days.clean %in% days.x ] )
+    }
+    return( list( x = x, modified = MODIFIED ) )
   }
-  return( list( x = x, modified = MODIFIED ) )
-}
 
 ##' @title Calculates the mode of a PDF
 ##' @description Calculates the mode of a PDF
@@ -624,16 +713,23 @@ mode <- function( x ){
 ##' 
 ##' @param x Either a time series of class \pkg{xts} or a list of
 ##'   them.
+##' @param mc.cores A numerical input specifying the number of cores
+##'   to use for the multi core application of the function (see
+##'   \code{\link[parallel]{detectCores}}). This functionality is only
+##'   available if the input is a list of different objects. If NULL,
+##'   the function will be calculated classically with only one core.
+##'   Default = NULL.
 ##'
 ##' @family ts
 ##' @export
 ##' @importFrom xts xts
 ##' @importFrom lubridate yday
+##' @importFrom parallel mclapply
 ##' 
 ##' @return Same class as the input
 ##' @author Philipp Mueller
 ##' 
-anomalies <- function( x ){
+anomalies <- function( x, mc.cores = NULL ){
   UseMethod( "anomalies" )
 }
 ##' @title Anomalies of a time series
@@ -646,17 +742,28 @@ anomalies <- function( x ){
 ##'   Uses the \code{\link[lubridate]{yday}}.
 ##' 
 ##' @param x List of time series of class \pkg{xts}
+##' @param mc.cores A numerical input specifying the number of cores
+##'   to use for the multi core application of the function (see
+##'   \code{\link[parallel]{detectCores}}). This functionality is only
+##'   available if the input is a list of different objects. If NULL,
+##'   the function will be calculated classically with only one core.
+##'   Default = NULL.
 ##'
 ##' @family ts
 ##' @export
 ##' @importFrom xts xts
 ##' @importFrom lubridate yday
+##' @importFrom parallel mclapply
 ##' 
 ##' @return Same class as the input
 ##' @author Philipp Mueller
 ##' 
-anomalies.list <- function( x ){
-  return( lapply( x, anomalies ) )
+anomalies.list <- function( x, mc.cores = NULL ){
+  if ( !is.null( mc.cores ) ){
+    return( mclapply( x, anomalies, mc.cores = mc.cores ) )
+  } else {
+    return( lapply( x, anomalies, mc.cores = mc.cores ) )
+  }
 }
 ##' @title Anomalies of a time series
 ##' @description Calculates the anomalies of an object of class
@@ -668,15 +775,22 @@ anomalies.list <- function( x ){
 ##'   Uses the \code{\link[lubridate]{yday}}.
 ##' 
 ##' @param x A time series of class \pkg{xts}
+##' @param mc.cores A numerical input specifying the number of cores
+##'   to use for the multi core application of the function (see
+##'   \code{\link[parallel]{detectCores}}). This functionality is only
+##'   available if the input is a list of different objects. If NULL,
+##'   the function will be calculated classically with only one core.
+##'   Default = NULL.
 ##'
 ##' @family ts
 ##' @export
 ##' @importFrom xts xts
 ##' @importFrom lubridate yday
+##' @importFrom parallel mclapply
 ##' 
 ##' @return Same class as the input
 ##' @author Philipp Mueller
-anomalies.xts <- anomalies.default <- function( x ){
+anomalies.xts <- anomalies.default <- function( x, mc.cores = NULL ){
   if ( !any( class( x ) == "xts" ) )
     stop( "Only for the class 'xts' the anomalies can be calculated" )
   ## base::ave seems to calculate takes a vector and a factor vector
@@ -684,8 +798,8 @@ anomalies.xts <- anomalies.default <- function( x ){
   ## sharing a factor and is placing the mean value in the element
   ## corresponding to the position of the original one.
   x.anomalies <- x - stats::ave( x, as.factor( yday( x ) ),
-                               FUN = function( xx )
-                                 mean( xx, na.rm = TRUE ) )
+                                FUN = function( xx )
+                                  mean( xx, na.rm = TRUE ) )
   return( x.anomalies )
 }
 ## End of ts.R